feat: move useAppTranslation hook to the hooks folder
