+++ src/components/screens/ChoosePasswordScreen.tsx
@@ -4,7 +4,7 @@ import Icon from "components/sds/Icon";
 import { Input } from "components/sds/Input";
 import { PASSWORD_MIN_LENGTH } from "config/constants";
 import { AUTH_STACK_ROUTES, AuthStackParamList } from "config/routes";
-import useAppTranslation from "helpers/useAppTranslation";
+import useAppTranslation from "hooks/useAppTranslation";
 import React, { useState } from "react";
 
 type ChoosePasswordScreenProps = {
