+++ src/components/screens/HistoryScreen.tsx
@@ -2,7 +2,7 @@ import RecoveryPhraseWarningBox from "components/RecoveryPhraseWarningBox";
 import { BaseLayout } from "components/layout/BaseLayout";
 import { THEME } from "config/theme";
 import { fs, px } from "helpers/dimensions";
-import useAppTranslation from "helpers/useAppTranslation";
+import useAppTranslation from "hooks/useAppTranslation";
 import React from "react";
 import styled from "styled-components/native";
